refactor(pagination): extract FilterToGridifyAdapter validation and normalization logic
Split validation and normalization concerns into dedicated partials. Keeps [GeneratedRegex] isolated in Normalization partial. Adapter main class focused on core mapping.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
